requirements.txt
ultralytics>=8.0.0
streamlit>=1.22.0
opencv-python>=4.7.0
pandas>=1.5.3
numpy>=1.24.3
shapely>=2.0.1
PyYAML>=6.0
Pillow>=9.5.0
matplotlib>=3.7.1
python-dotenv>=1.0.0 